requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64

accelerate=0.21.0=pypi_0
addict=2.4.0=pypi_0
aiohttp=3.8.5=py39h5eee18b_0
aiosignal=1.2.0=pyhd3eb1b0_0
albumentations=1.3.1=pypi_0
aliyun-python-sdk-core=2.13.36=pypi_0
aliyun-python-sdk-kms=2.16.1=pypi_0
antlr4-python3-runtime=4.9.3=pypi_0
appdirs=1.4.4=pypi_0
argparse=1.4.0=pypi_0
asttokens=2.0.5=pyhd3eb1b0_0
async-timeout=4.0.2=py39h06a4308_0
attrs=23.1.0=py39h06a4308_0
automat=22.10.0=pypi_0
aws-c-common=0.6.8=h5eee18b_1
aws-c-event-stream=0.1.6=h6a678d5_6
aws-checksums=0.1.11=h5eee18b_2
aws-sdk-cpp=1.8.185=h721c034_1
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.12.2=pypi_0
binutils_impl_linux-64=2.38=h2a08ee3_1
blas=1.0=mkl
blinker=1.6.2=py39h06a4308_0
boost-cpp=1.82.0=hdb19cb5_2
bottleneck=1.3.5=py39h7deecbd_0
brotli-python=1.0.9=py39h6a678d5_7
buildtools=1.0.6=pypi_0
bzip2=1.0.8=h7b6447c_0
c-ares=1.19.1=h5eee18b_0
ca-certificates=2023.08.22=h06a4308_0
cachetools=4.2.2=pyhd3eb1b0_0
causal-conv1d=1.1.3.post1=pypi_0
certifi=2023.7.22=py39h06a4308_0
cffi=1.15.1=py39h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=py39h06a4308_0
clip=1.0=pypi_0
colorama=0.4.6=pypi_0
comm=0.1.4=pypi_0
constantly=23.10.4=pypi_0
contourpy=1.1.0=pypi_0
crcmod=1.7=pypi_0
cryptography=41.0.3=py39hdda0065_0
datasets=2.12.0=py39h06a4308_0
debugpy=1.6.7=py39h6a678d5_0
decorator=5.1.1=pyhd3eb1b0_0
detail=4.0=pypi_0
diffusers=0.17.1=pypi_0
dill=0.3.6=py39h06a4308_0
docker-pycreds=0.4.0=pypi_0
docopt=0.6.2=pypi_0
einops=0.6.1=pypi_0
exceptiongroup=1.0.4=py39h06a4308_0
executing=0.8.3=pyhd3eb1b0_0
fairscale=0.4.0=pypi_0
ffmpeg=4.2.2=h20bf706_0
filelock=3.12.2=pypi_0
fonttools=4.42.1=pypi_0
freetype=2.12.1=h4a9f257_0
frozenlist=1.4.0=py39h5eee18b_0
fsspec=2023.6.0=pypi_0
ftfy=6.1.1=pypi_0
furl=2.1.3=pypi_0
fvcore=0.1.6=pypi_0
gcc=12.1.0=h9ea6d83_10
gcc_impl_linux-64=12.1.0=hea43390_17
gdown=4.7.1=pypi_0
gds-tools=1.7.1.12=0
gflags=2.2.2=he6710b0_0
giflib=5.2.1=h5eee18b_3
gitdb=4.0.10=pypi_0
gitpython=3.1.32=pypi_0
glog=0.5.0=h2531618_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
google-auth=2.22.0=py39h06a4308_0
google-auth-oauthlib=0.5.2=py39h06a4308_0
greenlet=3.0.3=pypi_0
grpc-cpp=1.48.2=he1ff14a_1
grpcio=1.48.2=py39he1ff14a_1
h5py=3.10.0=pypi_0
huggingface-hub=0.24.6=pypi_0
huggingface_hub=0.17.3=py39h06a4308_0
hyperlink=21.0.0=pypi_0
icu=73.1=h6a678d5_0
idna=3.4=py39h06a4308_0
imageio=2.31.1=pypi_0
importlib-metadata=6.8.0=pypi_0
importlib-resources=6.0.1=pypi_0
importlib_metadata=6.0.0=hd3eb1b0_0
incremental=22.10.0=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46306
iopath=0.1.10=pypi_0
ipykernel=6.25.0=py39h2f386ee_0
ipython=8.15.0=py39h06a4308_0
ipywidgets=8.1.1=pypi_0
jedi=0.18.1=py39h06a4308_1
jinja2=3.1.3=pypi_0
jmespath=0.10.0=pypi_0
joblib=1.2.0=py39h06a4308_0
jpeg=9e=h5eee18b_1
jupyter_client=8.6.0=py39h06a4308_0
jupyter_core=5.5.0=py39h06a4308_0
jupyterlab-widgets=3.0.9=pypi_0
kernel-headers_linux-64=2.6.32=he073ed8_15
kiwisolver=1.4.4=pypi_0
krb5=1.20.1=h143b758_1
lame=3.100=h7b6447c_0
lazy-loader=0.3=pypi_0
lcms2=2.12=h3be6417_0
mmcv=2.0.1=pypi_0
mmengine=0.8.4=pypi_0
model-index=0.1.11=pypi_0
multidict=6.0.2=py39h5eee18b_0
multiprocess=0.70.14=py39h06a4308_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=py39h06a4308_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=pypi_0
ninja=1.11.1.1=pypi_0
nsight-compute=2023.2.1.3=0
numexpr=2.8.7=py39h85018f9_0
numpy=1.26.0=py39h5f9d8c6_0
numpy-base=1.26.0=py39hb5e798b_0
oauthlib=3.2.2=py39h06a4308_0
ogb=1.3.6=pypi_0
omegaconf=2.3.0=pypi_0
open-clip-torch=2.20.0=pypi_0
opencv-python=4.8.0.76=pypi_0
opencv-python-headless=4.8.0.76=pypi_0
opendatalab=0.0.10=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openmim=0.3.9=pypi_0
openssl=3.0.12=h7f8727e_0
openxlab=0.0.19=pypi_0
orc=1.7.4=hb3bc3d3_1
ordered-set=4.1.0=pypi_0
orderedmultidict=1.0.1=pypi_0
oss2=2.17.0=pypi_0
outdated=0.2.2=pypi_0
packaging=23.1=py39h06a4308_0
pandas=2.1.1=py39h1128e8f_0
parso=0.8.3=pyhd3eb1b0_0
pathtools=0.1.2=pypi_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=10.0.1=py39ha6cbd5a_0
pip=23.3=py39h06a4308_0
platformdirs=3.10.0=py39h06a4308_0
portalocker=2.8.2=pypi_0
prettytable=3.8.0=pypi_0
prompt-toolkit=3.0.36=py39h06a4308_0
protobuf=3.20.3=py39h6a678d5_0
psutil=5.9.5=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pyarrow=11.0.0=py39h468efa6_1
pyasn1=0.4.8=pyhd3eb1b0_0
pyasn1-modules=0.2.8=py_0
pycocotools=2.0.7=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pycryptodome=3.18.0=pypi_0
pygments=2.16.1=pypi_0
pyjwt=2.4.0=py39h06a4308_0
pyopenssl=23.2.0=py39h06a4308_0
pyparsing=3.0.9=pypi_0
pysocks=1.7.1=py39h06a4308_0
python=3.9.17=h955ad1f_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-tzdata=2023.3=pyhd3eb1b0_0
python-xxhash=2.0.2=py39h5eee18b_1
pytorch=1.13.1=py3.9_cuda11.6_cudnn8.3.2_0
pytorch-cuda=11.6=h867d48c_1
pytorch-mutex=1.0=cuda
pytz=2023.3=pypi_0
pywavelets=1.4.1=pypi_0
pyyaml=6.0.1=py39h5eee18b_0
pyzmq=25.1.0=py39h6a678d5_0
qudida=0.0.4=pypi_0
re2=2022.04.01=h295c915_0
readline=8.2=h5eee18b_0
redo=2.0.4=pypi_0
regex=2023.10.3=py39h5eee18b_0
requests=2.28.2=pypi_0
requests-oauthlib=1.3.0=py_0
responses=0.13.3=pyhd3eb1b0_0
rich=13.4.2=pypi_0
rsa=4.7.2=pyhd3eb1b0_1
ruamel-yaml=0.18.5=pypi_0
ruamel-yaml-clib=0.2.8=pypi_0
sacremoses=0.0.43=pyhd3eb1b0_0
safetensors=0.3.2=pypi_0
scikit-image=0.21.0=pypi_0
scikit-learn=1.3.0=py39h1128e8f_0
scipy=1.11.3=py39h5f9d8c6_0
seaborn=0.13.0=pypi_0
sentencepiece=0.1.99=pypi_0
sentry-sdk=1.29.2=pypi_0
setproctitle=1.3.2=pypi_0
setuptools=60.2.0=pypi_0
shapely=2.0.1=pypi_0
simplejson=3.19.2=pypi_0
six=1.16.0=pyhd3eb1b0_1
smmap=5.0.0=pypi_0
snappy=1.1.9=h295c915_0
soupsieve=2.4.1=pypi_0
sqlalchemy=2.0.27=pypi_0
sqlite=3.41.2=h5eee18b_0
stack_data=0.2.0=pyhd3eb1b0_0
sysroot_linux-64=2.12=he073ed8_15
tabulate=0.9.0=pypi_0
tbb=2021.8.0=hdb19cb5_0
tensorboard=2.12.1=py39h06a4308_0
tensorboard-data-server=0.7.0=py39h52d8a92_0
tensorboard-plugin-wit=1.8.1=py39h06a4308_0
termcolor=2.3.0=pypi_0
terminaltables=3.1.10=pypi_0
threadpoolctl=2.2.0=pyh0d69192_0
tifffile=2023.8.12=pypi_0
timm=0.9.5=pypi_0
tk=8.6.12=h1ccaba5_0
tokenizers=0.13.3=py39h22610ee_0
tomli=2.0.1=pypi_0
torchvision=0.14.1=py39_cu116
tornado=6.3.3=py39h5eee18b_0
tqdm=4.66.1=pypi_0
traitlets=5.7.1=py39h06a4308_0
transformers=4.29.2=py39h06a4308_0
triton=2.1.0=pypi_0
twisted=23.10.0=pypi_0
typing-extensions=4.7.1=py39h06a4308_0
typing_extensions=4.7.1=py39h06a4308_0
tzdata=2023c=h04d1e81_0
urllib3=1.26.18=py39h06a4308_0
utf8proc=2.6.1=h27cfd23_0
wandb=0.15.8=pypi_0
wcwidth=0.2.6=pypi_0
werkzeug=2.2.3=py39h06a4308_0
wheel=0.41.2=py39h06a4308_0
widgetsnbextension=4.0.9=pypi_0
wilds=2.0.0=pypi_0
x264=1!157.20191217=h7b6447c_0
xformers=0.0.21=py39_cu11.7.1_pyt1.13.1
